+++ optimum/exporters/neuron/model_configs/__init__.py
@@ -1,0 +1,22 @@
+# coding=utf-8
+# Copyright 2023 The HuggingFace Team. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Model specific Neuron configurations."""
+
+from ....neuron.utils.import_utils import is_transformers_neuronx_available
+from .traced_configs import *
+
+
+if is_transformers_neuronx_available():
+    from .decoder_configs import *
